% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbitRE_PLNRE.R
\name{predict_ProbitRE_PLNRE}
\alias{predict_ProbitRE_PLNRE}
\title{Predictions of ProbitRE_PLNRE model on new sample}
\usage{
predict_ProbitRE_PLNRE(
  par,
  sel_form,
  out_form,
  data,
  offset_w_name = NULL,
  offset_x_name = NULL
)
}
\arguments{
\item{par}{Model estimates}

\item{sel_form}{Formula for selection equation, a Probit model with random effects}

\item{out_form}{Formula for outcome equation, a Poisson Lognormal model with random effects}

\item{data}{Input data, a data.frame object}

\item{offset_w_name}{Offset variables in selection equation, if any.}

\item{offset_x_name}{Offset variables in outcome equation, if any.}
}
\value{
A list with three sets of predictions
\itemize{
\item prob: Predicted probability to participate
\item outcome: Predicted potential outcome
\item actual_outcome: Predicted actual outcome
}
}
\description{
Predictions of ProbitRE_PLNRE model on new sample. Please make sure the factor variables in the test data do not have levels not shown in the training data.
}
